+++ provision/provision.sh
@@ -60,7 +60,6 @@ apt_package_check_list=(
 	php5-curl
 	php-pear
 	php5-gd
-	php-apc
 
 	# memcached is made available for object caching
 	memcached
@@ -199,7 +198,7 @@ if [[ $ping_result == *bytes?from* ]]; then
 		chmod +x composer.phar
 		mv composer.phar /usr/local/bin/composer
 
-		COMPOSER_HOME=/usr/local/src/composer composer -q global require --no-update phpunit/phpunit:3.7.*
+		COMPOSER_HOME=/usr/local/src/composer composer -q global require --no-update phpunit/phpunit:4.0.*
 		COMPOSER_HOME=/usr/local/src/composer composer -q global require --no-update phpunit/php-invoker:1.1.*
 		COMPOSER_HOME=/usr/local/src/composer composer -q global require --no-update mockery/mockery:0.8.*
 		COMPOSER_HOME=/usr/local/src/composer composer -q global require --no-update d11wtq/boris:v1.0.2
@@ -264,7 +263,15 @@ fi
 # Used to to ensure proper services are started on `vagrant up`
 cp /srv/config/init/vvv-start.conf /etc/init/vvv-start.conf
 
-echo " * /srv/config/init/vvv-start.conf               -> /etc/init/vvv-start.conf"
+cp /srv/config/php5-apache2-config/php-custom.ini /etc/php5/apache2/conf.d/php-custom.ini
+cp /srv/config/php5-apache2-config/opcache.ini /etc/php5/apache2/conf.d/opcache.ini
+cp /srv/config/php5-apache2-config/xdebug.ini /etc/php5/apache2/conf.d/xdebug.ini
+
+echo " * /srv/config/php5-apache2-config/opcache.ini    -> /etc/php5/apache2/conf.d/opcache.ini"
+echo " * /srv/config/php5-apache2-config/php-custom.ini -> /etc/php5/apache2/conf.d/php-custom.ini"
+echo " * /srv/config/php5-apache2-config/xdebug.ini     -> /etc/php5/apache2/conf.d/xdebug.ini"
+
+echo " * /srv/config/init/vvv-start.conf                -> /etc/init/vvv-start.conf"
 
 # Copy memcached configuration from local
 cp /srv/config/memcached-config/memcached.conf /etc/memcached.conf
@@ -373,6 +380,18 @@ if [[ $ping_result == *bytes?from* ]]; then
 		echo "phpMemcachedAdmin already installed."
 	fi
 
+	# Checkout Opcache Status to provide a dashboard for viewing statistics
+	# about PHP's built in opcache.
+	if [[ ! -d /srv/www/default/opcache-status ]]; then
+		echo -e "\nDownloading Opcache Status, see https://github.com/rlerdorf/opcache-status/"
+		cd /srv/www/default
+		git clone https://github.com/rlerdorf/opcache-status.git opcache-status
+	else
+		echo -e "\nUpdating Opcache Status"
+		cd /srv/www/default/opcache-status
+		git pull --rebase origin master
+	fi
+
 	# Webgrind install (for viewing callgrind/cachegrind files produced by
 	# xdebug profiler)
 	if [[ ! -d /srv/www/default/webgrind ]]; then
@@ -462,7 +481,15 @@ PHP
 	else
 		echo "Updating WordPress develop..."
 		cd /srv/www/wordpress-develop/
-		svn up
+		if [[ -e .svn ]]; then
+			svn up
+		else
+			if [[ $(git rev-parse --abbrev-ref HEAD) == 'master' ]]; then
+				git pull --no-edit git://develop.git.wordpress.org/ master
+			else
+				echo "Skip auto git pull on develop.git.wordpress.org since not on master branch"
+			fi
+		fi
 		npm install &>/dev/null
 	fi